fix(nodesystem): paste nodes to right position

+++ packages/nodesystem/src/view/NodeSystemView.ts
@@ -1,3 +1,4 @@
+import './NodeSystemView.scss';
 import { EventEmitter, throttle } from '@plantarium/helpers';
 import { createPanZoom } from '../helpers';
 import type Node from '../model/Node';
@@ -7,7 +8,6 @@ import type NodeSystem from '../model/NodeSystem';
 import AddMenu from './AddMenu';
 import BoxSelection from './BoxSelection';
 import FloatingConnectionView from './FloatingConnectionView';
-import './NodeSystemView.scss';
 import ColorStore from './socketColorStore';
 
 export default class NodeSystemView extends EventEmitter {
@@ -334,9 +334,6 @@ export default class NodeSystemView extends EventEmitter {
 
     const { x, y } = this.projectWindowToLocal(this.rmx, this.rmy);
 
-    this.mx = x;
-    this.my = y;
-
     this.selectedNodesDown = this.selectedNodes.map((_n) => [
       _n.view.x,
       _n.view.y,
@@ -396,13 +393,21 @@ export default class NodeSystemView extends EventEmitter {
         break;
       case 'c':
         if (shiftKey && ctrlKey) {
-          localStorage.clear();
-          window.location.reload();
+          if (window.confirm("Clear Storage?")) {
+            localStorage.clear();
+            window.location.reload();
+          }
         } else if (ctrlKey) {
           const s = this.selectedNodes.splice(0);
           if (this.activeNode && !s.includes(this.activeNode))
             s.push(this.activeNode);
-          this.clipboard = s.map((n) => n.deserialize());
+          this.clipboard = s.map((n) => n.deserialize()).map(n => {
+            n.attributes.pos.x -= this.mx;
+            n.attributes.pos.y -= this.my;
+            console.table({ ...n.attributes.pos, mx: this.mx, my: this.my })
+            return n
+          });
+          console.log(this.clipboard)
         } else if (this.selectedNodes.length && this.activeNode) {
           this.selectedNodes[0].connectTo(this.activeNode);
         }
@@ -458,29 +463,19 @@ export default class NodeSystemView extends EventEmitter {
       // v
       case 'v':
         if (ctrlKey) {
-          const sorted = this.clipboard.sort((a, b) => {
-            const { pos: { x: x1 = 0, y: y1 = 0 } = {} } = a.attributes;
-            const { pos: { x: x2 = 0, y: y2 = 0 } = {} } = b.attributes;
-            return x1 + y1 < x2 + y2 ? -1 : 1;
-          });
 
-          let { pos: { x: offsetX = 0, y: offsetY = 0 } = {} } =
-            sorted[0].attributes;
-
-          const { x: mx, y: my } = this.projectWindowToLocal(this.mx, this.my);
-
-          offsetX -= mx;
-          offsetY -= my;
 
           this.clipboard
-            .map((prop) => {
-              const { pos: { x = 0, y = 0 } = {} } = prop.attributes;
+            .map((node) => {
+              const { pos: { x = 0, y = 0 } = {} } = node.attributes;
 
-              prop.attributes.pos = {
-                x: x - offsetX,
-                y: y - offsetY,
+              node.attributes.pos = {
+                x: x + this.mx,
+                y: y + this.my,
               };
-              return prop;
+
+              console.table({ x, y, mx: this.mx, my: this.my })
+              return node;
             })
             .forEach((c) => this.system.createNode(c));
         }
